Initiall support for allocated places
