+++ app/views/_includes/summary-cards/programme-details.html
@@ -1,7 +1,10 @@
 {% set sectionIsRequired = record | requiresSection("programmeDetails") %}
 
 {% if sectionIsRequired %}
-  {# {% if record.route == 'Assessment Only' %} #}
+  {% if record.route == 'Assessment Only' %}
     {% include "_includes/summary-cards/programme-details/assessment-only.html" %}
-  {# {% endif %} #}
+  {% endif %}
+  {% if record.route == 'Provider-led' %}
+    {% include "_includes/summary-cards/programme-details/provider-led.html" %}
+  {% endif %}
 {% endif %}
