+++ app/views/new-record/programme-details/allocated-place.html
@@ -1,0 +1,10 @@
+{% extends "_templates/_new-record.html" %}
+
+{% set pageHeading = "This course has allocated places" %}
+
+{% set formAction = "./confirm" %}
+
+{% block formContent %}
+  {% include "_includes/forms/programme-details/allocated-place.html" %}
+{% endblock %}
+
